requirements.txt
chromadb>=0.5.0
openai>=1.40.0
python-dotenv>=1.0.1
